requirements.txt
langchain
langchain-community 
langchain-openai
langchain-experimental
python-dotenv
langchain_ollama
reportlab
google-generativeai
PyPDF2
pdfplumber
pytesseract
Pillow
pdf2image
poppler-utils